Catch exceptions on password resets creation
It happened sometimes that there was a misconfiguration on the mailer
values and then the passwords controller raised exceptions on password
reset creation. This is bad because the user will face an empty page,
and the logs won't reflect a thing.

With this commit we will catch these exceptions and properly log it.
Furthermore, the user will be redirected to the original page.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>