packaging: do not modify the Gemfile file
Since the very beginning we've added some code blocks in the Gemfile
that run conditionally. This was bad practice but it worked for us. This
is no longer the case: current builds fail because bundle exits with a
non-zero code after checking that the Gemfile.lock produced is not the
same.

Thus, let's do things properly: use groups instead. I've done the
following:

- Removed the `packaging?` method since it's not needed anymore.
- Added an `assets` group that will be excluded when installing gems.
- Explicitely exclude `development` and `tests` instead of doing it on
  an ugly `unless` block.
- Removed the conditional on the `puma` gem. People are expected to run
  Portus with puma always.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>