packaging: changed how we build dependencies
This commit is a follow-up of bd383fba329b6256fdd0bb6d1a07fa98b12aeff0,
and it adapts all the packaging scripts on the idea that the Gemfile and
the Gemfile.lock files cannot be touched. After this commit, the portus
RPM builds again and it's cleaner than ever.

First of all, the `make_spec.sh` has been heavily modified, while
polishing some rough edges (e.g. proper debug functions). It will
require a specific bundler version, because now it will be taken from
the gem requirements instead of the `spec.in` file. The biggest change
is to only use `bundler` for anything related to gems. To list the gems
to be added as dependencies we are using now `bundle show`, which has
the knowledge of gem groups as defined in the previous commit. This way,
without modifying the Gemfile.lock file, we can properly filter the
dependencies. As a side-effect, this simplifies quite a lot the script.

As for the `portus.spec.in`, I have updated it so it uses gem groups,
and it will compile assets after fetching all gems (since it's the proper
order). Finally, I also added a couple of lines to remove all the
unneeded files/directories coming from vendored gems.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>