examples: use the PORTUS_INIT_COMMAND variable
This is a regression from ced9b46a9064a7c3e4374e427f3201a200db8c0a, in
which the entrypoint was set manually. This is a bad idea because the
Docker image has a special environment (e.g. gems deployed in production
in the `vendor` directory), and the default entrypoint already sets the
proper environment variables. Thus, we should only instruct the default
init process to run `bin/background` in the end.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>