environments: do not migrate by default
The migration part is now expected to be done by the administrator. This
is done automatically by the init script in containerized deployments.

This was done because now we had at least two portus processes accessing
the DB and trying to migrate the DB, which caused quite some issues.
Moreover, notice that the previous workflow (e.g. automatically
migrating on production) was simply not possible in orchestration
solutions like Docker Swarm or Kubernetes.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>