js: introduced unit testing
We are introducing the ability of writing javascript unit tests. Until now,
this wasn't possible.

To achieve that we decided to go with the mocha + webpack approach that
uses the `mocha-webpack` package. As the assertion library we choose
`expect`.

To run the tests you can simply type `yarn test`. And if you want the to check
the coverage, you can type `yarn coverage`.

There still stuff to be improved. The `.vue` files coverage does not work
properly and this is something that needs investigation.