Upgraded to ruby 2.5
Apparently in SLE 15 (and so Tumbleweed too) the default ruby will be
2.5 instead of 2.4 as we were told. This commit adds support for this
version and sets it as the default (so we anticipate the release).

This change involved a major change: devise had to be upgraded from
3.5.0 to 4.4.0. Devise was quite old and only in 4.4.0 it received some
fixes for this ruby version. Hence, this commit also includes quite some
changes to accomodate this upgrade.

Finally, I've also fixed some minor issues (e.g. the order of the errors
controller in the routes.rb file), and I've upgraded the rubocop version
so it catches the latests things from ruby 2.5 as well.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>