packaging: manually set the bundler version
The bundler version is determined by the one being used. This is the
ideal situation, but it turns out it's basically impossible to guarantee
that on Travis CI (we were using an ugly hack that does not work anymore
because they've updated stuff...).

So, the idea is that the bundler version being used doesn't matter that
much, and we are already using it on development, so we'll set it
directly when creating the spec file.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>