OpenID Connect & Active Session Storage
Integrated with `https://github.com/hhorikawa/omniauth-openid-connect` to provide a simple OpenID Connect authorization. There was a few other bugs that I ran into.

when storing the Omniauth respose into the session there was mixed code. One was treating the object as a hash and another was treating it as a omniauth object. There was only one reference to the omniauth object so I just converted the object to an hash and then changed the one reference to the omniauth object to be a hash reference.

With the pull of the omniauth-openid-connect being from github the docker file used to create the development image needs git to also be installed.

Added social link for OpenID Connect. The verbiage for hte link was too long so I made it say ODIC on the social icon. I do think we need to revist how this is generated. With the addition of OpenID Connect and OpenID there maybe a need to specify several social icons with the same type. Both the social icons and the config will need to change to accomodate this. I opted to not include this change because it could break current installations when upgrading.

Session storage - I ran into a limitation where the information stored inside of the session was exceeding 4kb. This is because currently portus uses cookie based session storage which has a limitation of 4kb. unfortunatly openid requries 4.1kb for the tokens and certs. I switched to the other common session storage to be `active_record_store` which is a database driven session storage. This really should of been done sense the first implementation of omniauth as there is some critical information in that data that could be abused if stored in a cookie. I am also surprised that the 4kb limit hasnt been broken for other things in Portus.

Added simple spec for openid_connect. Since i am not sure how to run the tests locally I think it will fail when travis-ci runs it.