Restructured pages and simplified home page
First of all, I have upgraded this to the latest Jekyll. The version we
were using was really old, and it failed on Travis. Moreover, we are
now taking advantage of Jekyll's SASS support. This comes out of the
box since Jekyll 3.0, and it greatly simplifies development environment.

Moreover, I have dropped Gulp, which included lots of dependencies.
Instead, I've added a package.json file, which is then picked up by
Yarn. This results in a much lighter node_modules directory. Moreover,
I've also added webpack for the javascript. The removal of Gulp has also
involved a re-structure of the assets directories.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>