Updated the deployment strategies
A new page has been added on how to deploy Portus. In this page we will
offer a complete guide on how to deploy Portus, and we will hint some of
the recommendations.

Moreover, a link has been added in the main page, and outdated pages
have been either removed or vastly updated.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>