docs: added documentation for the API
I've created some small Jekyll plugins to overcome some flaws from the liquid
markup language... The code for the API documentation was inspired by
https://github.com/jexhson/jekyll-swagger, but I ultimately changed it so much
that it does not resemble in any way that plugin.

Moreover, because of the creation of these plugins, rubocop was getting noisier,
so I adapted the configuration from the `master` branch into this one.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>