background: mark failed scans as re-schedulable
After a security scanning iteration, some tags might not be in a
:scan_done state. This is possible in situtations like:

- Clair was unavailable for some time because of a hiccup.
- There's an actual bug on the security scanning.

In these cases, we will check in the end of the iteration whether all
tags have been marked as scan. If this is not the case, then we will log
a warning and mark the failed tags as :scan_none (so in the next
iteration they can be picked up and hopefully be scanned properly this
time).

See #1649

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>