sync: added sync-strategy as a config value
We have historically struggled with the synchronization of the Registry.
This was developed at first because the following situation could
happen:

1. User pushes image:tag.
2. Portus authorizes the transaction.
3. Registry sends the event to Portus with all the data, but Portus is
   unavailable.
4. Portus is up again but it missed the registry event.

This is not the case anymore, because all the things that could make
Portus unavailable have been either fixed or moved into the background
process. Thus, this problem can be ignored in most cases. Moreover, this
feature is quite dangerous because a bug on this code can make Portus
wipe all the repositories, and historically "funny" issues have
happened.

This commit builds upon #1631 where background tasks can be disabled.
Moreover, it also adds a `sync-strategy` option which is available when
users enable this feature. This option has four possible values:

- `update-delete`: the same behavior we've had up until now.
- `update`: similar to `update-delete` but this one will not delete
  repositories which no longer exist on the registry. This is useful for
  users which don't trust the risky `update-delete` strategy.
- `on-start`: execute `update-delete` only once (on start).
- `initial`: execute `update-delete` only once on start and
  only if the registry is empty. This is the default strategy since it
  might be convenient for new users that have already a running
  registry.

This commit does not fix all the issues mentioned in #1599, but it
hopefully does the trick for most users.

Fixes #1650
Fixes #1664
See #1599
Depends on #1631

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>