feature: contributors allowed to delete repositories/tags
With the variety usage of Portus, the need of giving contributors more
responsibilities arose. So far only admin were able to delete repositories
and tags if the capability were enabled. This patch extends the feature to
also allow contributors to be able delete resources.

We also took the opportunity to migrate and add two endpoints [0] [1] the
API regarding this topic. Users are now able to invoke the API to perform
those actions if they want to.

[0] DELETE /api/v1/repositories/:id
[1] DELETE /api/v1/tags/:id

Signed-off-by: Vítor Avelino <vavelino@suse.com>