teams: improved team creation form with owner
If you are a Portus admin, you might want to create various teams with
its respective owners. So far the path to achieve that was tedious and
not friendly. Admin had to create a team, add a member as owner and
remove himself as ownere later on.

With this patch, we are unifying the admin team page with the normal
team page and adding the ability to assign a different owner for a
team.

API also had an impact with this change. The GET /api/v1/teams endpoint
will expose all the accessible teams if the user is an admin. Very
similar to what you get when navigating through the UI.

Signed-off-by: Vítor Avelino <vavelino@suse.com>