security: fetch the manifest more safely
In order to fetch vulnerabilities, the security scanner first has to
gather the layers to be inspected. In our case, we fetch this
information from the image manifest as given by the Registry.

This commit adds some code so the fetching of this manifest is safer.
For example, it will not freak out if a timeout was reached when
requesting the manifest from the registry.

Fixes #1743

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>