lib: be explicit on the exceptions to be rescued
In the past we have had some problems when it comes to unhandled
exceptions. This commit fixes the opposite problem: having too broad
exceptions. That is, this commit clarifies which exact exceptions might
be raised from the different registry methods and updates the places
where these methods are used.

Some tests had to be corrected because they were abusing some code
paths that were rescuing StandardError.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>