namespaces: unified admin page with regular page
We got rid of the admin namespaces page and unified the vision of the
admin on the regular namespace page. So, if the user is an admin, they
will be seeing all of the namespaces and will be able to manage all of
them. If the user is a regular one, everything stays the same.

API is also affected by this change.

One improvement we did was to separate the other namespaces the user had
access to without being through membership in a different panel. This
will help the user understand better how they have access to those
namespaces.

Signed-off-by: Vítor Avelino <vavelino@suse.com>