repositories: splitted repositories into different panels
All the repositories were mixed together without distinction from the
repositories accessible via team membership or via visibility. Of
course, if the user is an admin, all the repositories would be listed.
Anyways, a nice enhancement to have this separation and create a sense
of semantic in the UI level.

Signed-off-by: Vítor Avelino <vavelino@suse.com>