dashboard: unified admin page with regular page
We got rid of the admin dashboard page. We finaly complete the unification
of admin tasks/vision on the regular pages. If the user is an admin,
they will see more things on the page now.

In this case, the dashboard page will show similar components to what the
user used to see on the admin dashboard page. The main difference is that
now we have tabs grouping the types of recent activities.

If the user is a regular one, nothing changes and everything stays the same.

Signed-off-by: Vítor Avelino <vavelino@suse.com>