sync: do not remove repositories on some errors
This commit is more explicit on what to do when a catalog error happens.
In this case, if a plain catalog error happens (fetching the list of
repositories and their manifests), then nothing will be done. If
fetching the catalog worked, and fetching the list of tags then suddenly
fails, before this commit that repo would've get nuked because it
would've get a repository with no tags.

This commit instructs to skip that repository when considering whether
to remove or not some repositories that are found to be dangling.

See #1293
See #1599

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>