Fixed hostname copied to clipboard on tags
The internal hostname was being referring as the hostname to be copied
to the clipboard when the user clicked on the labeled tags. This patch
fixes that by checking if external hostname is set or not. If set,
external hostname will be used instead of internal hostname.

This change affects the API directly because the repository entity is
now using the #reacheable_hostname method that tried to prioritize the
external hostname if set.

Signed-off-by: Vítor Avelino <vavelino@suse.com>