ui: migrated team and namespace update to api
With the implementation of update endpoints in the API, client side can
now take advantage of that and refactor some legacy code.

One major change that was the refactoring and migrating to vue components
of the team details panel. And a minor one of using new namespace update
endpoint in the namespace edit form.

Legacy js components of teams module were also removed.

Signed-off-by: Vítor Avelino <vavelino@suse.com>