js: replaced typeahead.js w/ vue-multiselect
Typeahead.js is not maintained anymore for a few years and it doesn't
work with modern js. After jQuery upgrade, we had to move jQuery out of
package.json because of typeahead.

In this patch typeahead.js was removed and replaced by vue-multiselect
and jQuery is back to package.json.

Signed-off-by: Vítor Avelino <vavelino@suse.com>