ui: added visibility to namespace edit form
The visibility change was only possible via namespaces table showed in
the namepace#index page. User can now change the visibility while
editing a namespace.

A <visibility-chooser> component was created to abstract most of the
visibilities options. <namespace-visibility> evolved to be a wrapper of
the previous component that can be used in standalone to change a
namespace remotely.

Signed-off-by: Vítor Avelino <vavelino@suse.com>