health: catch all exceptions for registries
The fact that `RegistryClient.reachable?` might raise an exception is
fine in some situations on the API, but it prevents us from knowing from
other errors, and it also makes things harder in other places of the
code that rely on this check. Hence, let's rescue this exception and
return its value properly.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>