test: fixed git-validation in travis
Travis apparently does not consider the whole git log, so when trying to
start git-validation from an old entry, it will fail. The solution is to
simply use the target branch if available, so in PRs we check
git-validation only for new commits.

I've also removed the quiet flag because after this change the output
shouldn't be that long.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>