[packaging] fix gem packaging
Instead of adding all the gems as sources, we will use the cpio
strategy, which does not need to change the spec file or, more
important, does not need us to add/remove the .gem files into the
package every time there is a change in the Gemfile.lock file.

However, using this strategy the gems are copied on another path, and we
need to fix this.

Then, we will need to adapt the _service file in obs to have these
sections, to download the Gemfile and Gemfile.lock, and to create the
cpio with the gems:

  <service name="download_url">
    <param name="protocol">https</param>
    <param name="host">raw.githubusercontent.com</param>
    <param name="path">/SUSE/Portus/master/Gemfile</param>
  </service>
  <service name="download_url">
    <param name="protocol">https</param>
    <param name="host">raw.githubusercontent.com</param>
    <param name="path">/SUSE/Portus/master/Gemfile.lock</param>
  </service>
  <service name="bundle_gems">
    <param name="strategy">cpio</param>
  </service>

A part from that, we need to add bundler-*.gem as a SOURCE in the spec
file, since the obs-service-bundle_gems does not add this gem into the
cpio.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>