ui: fixed repositories performance view issue
Previously all the repositories were fetched and that caused a
performance issue because of the amount of tags that could be associated
to the repositories.

Instead of fetching all the repositories at once, we introduced
pagination to the repositories#index page.

Signed-off-by: Vítor Avelino <vavelino@suse.com>

(cherry picked from commit ebba53f88ec39a0361465f6fb30024fb61d6a5ee)