ldap: automatically set admin according to filter
Introduced the new `ldap.admin_base` filter, which allows users to set a
filter for Portus administrators. When a new user logs in, it will
automatically set the `admin` flag depending on whether it was found on
this filter or not. If this configuration option is not set (default),
then this whole workflow will be ignored.

Fixes #1298

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>