ldap: prevent to create users colliding with LDAP
Administrators can create users by using the API or by using the admin
form. This commit forbids the collision of user creation with existing
LDAP users. This is done to avoid situations in which a user is created
inconsistently (e.g. you create an existing user and assign it a
different password than the version of the LDAP server). We should
prevent these inconsistencies if we want to allow the mixture of LDAP
and DB users.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>