ui: improved namespace transfer workflow
Preivously the only way of transferring a namespace was through the edit
namespace form. It wasn't that obivous.

A dedicated transfer button was added to the details panel header that
opens a dedicated modal to transfer when clicked.

As a plus, the style of edit namespace toggle was changed from link to
button. "namespace" redundant word was removed and delete button was
moved to the panel header.

Signed-off-by: Vítor Avelino <vavelino@suse.com>