ui: repositories#show content reorganization
Repositories#show page had its content finally reorganized. Sections
were organized into tabs. We had 'Tags' and 'Comments' in the same page
that turned into tabs and we've created a new tab named 'Overview'.

The 'Overview' tab now contains a description that is editable. This is
a new feature that will help users to provide a more general information
about the repository if necessary.

The workflow stays the same for the rest.

Signed-off-by: Vítor Avelino <vavelino@suse.com>