packaging: fixed wrong link on public_activity
Apparently there is a soft link on a test of public_activity, and OBS
build doesn't quite like them. So, this commit simply nukes the `test`
directory of this gem (which is, for our purposes, is never really
needed).

This has already been tried on OBS directly and it worked.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>