Removed unneeded directory
The `deploy` directory was there since back then when we didn't have a
clear deployment strategy. Since we moved to a container-based
deployment, this doesn't make much sense. Moreover, even if you are
deploying portus bare-metal, we never updated this Chef recipe so maybe
it's totally out-dated. Last but not least, I don't think such files
should be included into this repository.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>