fixed deletion for registry 2.7
The latest release of docker registry changed the way of interacting
with the jwt issuer and now requires the explicit "delete" permission
instead if "*" permission.

We've just added the proper method to namespace policy and allowed the
request to passthrough with that.

Signed-off-by: Vítor Avelino <vavelino@suse.com>
(cherry picked from commit c788e363215553f0f4947ed3a202edaa6fd2ae00)