ui: improved sidebar highlight
Previously the highlight was only happening when hovering and when the
`#index` pages where selected. When the user selected a page or were
navigating in a sub-page (e.g. /repositories/1), the highlight
disappeared.

With this patch we are enforcing the highlight in every single page the
user visits. This will give more context for the user.

Signed-off-by: Vítor Avelino <vavelino@suse.com>