Upgraded to ruby 2.5.3
Three patch-level releases have occurred since we last updated ruby, and some of
them include CVE fixes. So, it's important to upgraded to the latest patch-level
release.

Moreover, I've also upgraded rubocop, just to be sure that we are on the latest
code style guides from the community as well.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>